Change game structure to use phaser scenes
